Amend the default .htaccess rules we generate to include a final line that rewrites the HTTP_AUTHORIZATION header. In php-cgi configurations this header is not passed from the CGI handler to the destination script by default, so it is not available to PHP.
This rewrite tells Apache to create an environment variable called 'HTTP_AUTHORIZATION' and set it to the contents of %{HTTP:Authorization}, the Apache syntax for the HTTP Authorization header.

Since the CGI handler passes in all Apache environment variables, this rule emulates the default behavior of prepending 'HTTP_' to all HTTP headers. The CGI handler will not hand in the 'Authorization' header by default, so this gives us the same result by working around the dubious "security" restriction.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@4592 653ae4dd-d31e-0410-96ef-6bf7bf53c507