Revamp the Posts::get vocabulary-selection code so we don't fetch each individual term and instead just build a SQL string with the text values we're handed.
There are a few reasons for this. Most notably, the prior implementation would silently discard any terms that didn't exist, which means you would always get back an un-limited set of posts when trying to fetch for a non-existent term. This fixes #1369 as it simply plugs the term provided into the SQL query, which would never return a result.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@4748 653ae4dd-d31e-0410-96ef-6bf7bf53c507