THIS COMMIT WILL BREAK YOUR THEMES. Themes now require <class>Classname</class> in the theme XML that matches the class name in the theme.php. Core themes have been updated, and any other theme that has not been updated as such will be marked legacy. Defining THEME_CLASS in theme.php is not longer necessary and in fact will do nothing, so it has been removed from core themes.
Theme class names need to be unique, at least between the theme being deactivated and the one being activated.

All of this is to make r4852 work (to close ticket #1258) so that themes (and plugins) can react to the activation and deactivation of themes. For example, a theme can now register action_theme_activated() to create some content or blocks at activation.

For the time being refer to r4852, until the documentation is updated with these hooks.


git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@4875 653ae4dd-d31e-0410-96ef-6bf7bf53c507